+++ change-log.md
@@ -12,6 +12,7 @@
 - add: TestSuiteRunners add the TestConfiguration to the TestResult.
 - add: JUnitReportRenderer uses the TestConfiguration name as the test name
 - add: JUnitReportRenderer uses the TestSuite id as the class name
+- improve: VariableListContextInitializeConfiguration takes a ValueSourceConfiguration for the list id, instead of a contstant string.  
 - fix: OpenBrowser step should not propogate an exception during test resource cleanup if WebDriver.quit() throws an exception.
 - fix: BaseMuseTest always returns a result, even when a serious exception is thrown.
 - fix: encode jUnit report XML in UTF-8